# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VS Furniture API",
        "description": "API documentation for VS Furniture Laravel Backend",
        "contact": {
            "name": "VS Furniture",
            "email": "admin@vsfurniture.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/contact": {
            "get": {
                "tags": [
                    "Contact"
                ],
                "summary": "Get all contact messages (admin only)",
                "operationId": "fc65a9ced8387969bc5b37c8eeafb9db",
                "responses": {
                    "200": {
                        "description": "Messages retrieved successfully"
                    },
                    "403": {
                        "description": "Unauthorized - Admin access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contact"
                ],
                "summary": "Send contact message",
                "operationId": "3356e812494f1b29157eeb3640497bf6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "subject",
                                    "message"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "subject": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message sent successfully"
                    }
                }
            }
        },
        "/contact/{contactMessage}": {
            "get": {
                "tags": [
                    "Contact"
                ],
                "summary": "Get single contact message (admin only)",
                "operationId": "af044d94d0734682bcd1863e47d89abf",
                "parameters": [
                    {
                        "name": "contactMessage",
                        "in": "path",
                        "description": "Contact Message ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Contact"
                ],
                "summary": "Delete contact message (admin only)",
                "operationId": "2a632490950310becea5a4bb4a1e4a70",
                "parameters": [
                    {
                        "name": "contactMessage",
                        "in": "path",
                        "description": "Contact Message ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/contact/{contactMessage}/status": {
            "patch": {
                "tags": [
                    "Contact"
                ],
                "summary": "Update message status (admin only)",
                "operationId": "b637ed8e08e573c5a4bc1329933c27da",
                "parameters": [
                    {
                        "name": "contactMessage",
                        "in": "path",
                        "description": "Contact Message ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "unread",
                                            "read",
                                            "replied"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message status updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/contact/stats/overview": {
            "get": {
                "tags": [
                    "Contact"
                ],
                "summary": "Get contact message statistics (admin only)",
                "operationId": "ebc93a9039a049aa43bc029ee8eaf1f5",
                "responses": {
                    "200": {
                        "description": "Contact statistics retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}